pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aa-moonmaster"
version = "0.1.17"
description = "Unified moon mining & Metenox management for Alliance Auth — survey database, Athanor extractions, Metenox fuel/goo tracking, and real-time profitability with tax support."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Thrainkrilleve" },
]
keywords = ["alliance auth", "eve online", "moon mining", "metenox", "industry"]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "allianceauth>=4.0.0",
    "django-esi>=4.0.0",
    "django-eveuniverse>=1.0.0",
    "requests>=2.28.0",
]

[project.urls]
Homepage = "https://github.com/Thrainkrilleve/aa-moonmaster"
Repository = "https://github.com/Thrainkrilleve/aa-moonmaster"
"Bug Tracker" = "https://github.com/Thrainkrilleve/aa-moonmaster/issues"

[tool.setuptools.packages.find]
include = ["moonmaster*"]

[tool.setuptools.package-data]
moonmaster = [
    "migrations/*.py",
    "templates/moonmaster/*.html",
    "templates/moonmaster/**/*.html",
    "static/moonmaster/css/*.css",
    "static/moonmaster/js/*.js",
    "locale/*/LC_MESSAGES/*.po",
    "locale/*/LC_MESSAGES/*.mo",
]
